Use a prebuilt webview APK
Provide a prebuilt binary of webview, replacing the version built from
source in frameworks/webview/chromium/Android.mk. The cleanspec will
delete any old libwebviewchromium.so installed systemwide and any
symlinks to it that may have been in /system/app/webview/lib, as the
prebuilt's library is directly stored in the webview lib directory
instead of in /system/lib.